database_retrieval/requirements.txt
openai==1.66.3
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.43.2
